Add files via upload
With this update, it is possible to read the RadioText information on the receiver's LCD display, adapting the table of characters on the display with that of the RDS, thus correctly presenting the accented letters. The texts that are sent through the serial port are converted to UTF8, also correcting the presentation defects that occurred when displaying the RDS texts on mobile devices.